Added game logic
working on finding a way to determine if someone has won the game.
I'm currently pushing x and o values into an array targeting the
ids. Then I'm trying to check to see if a winning combination of
elements within the array exist to determine the winner. Not sure
how to do it yet.