cleaned code, added style, made message box functional
message box now displays who won, if signing
in and up and changing password was successful, and the number of games
played by the registered user. I turned long pieces of code into funtions.
I added a bit of css and a description of the game.  finally I completed
the README.md file.